use span for bytes reader and buffer

+++ gojo/syscall/__init__.mojo
@@ -48,7 +48,7 @@ alias c_short = Int16
 alias c_ushort = UInt16
 alias c_int = Int32
 alias c_uint = UInt32
-alias c_long = Int64
+alias c_long = Int
 alias c_ulong = UInt64
 alias c_float = Float32
 alias c_double = Float64
@@ -57,8 +57,8 @@ alias c_double = Float64
 alias c_size_t = Int
 alias c_ssize_t = Int
 
-alias ptrdiff_t = Int64
-alias intptr_t = Int64
+alias ptrdiff_t = Int
+alias intptr_t = Int
 alias uintptr_t = UInt64
